Fixing dtype specification and device map for multi-gpu setups

+++ examples/quickstart.py
@@ -2,7 +2,7 @@
 import torch
 from neuron_steer import NeuronSteerer
 
-steerer = NeuronSteerer("meta-llama/Llama-3.1-8B-Instruct", dtype=torch.bfloat16)
+steerer = NeuronSteerer("meta-llama/Llama-3.1-8B-Instruct")
 
 # Discover the circuit for "say a capital city"
 circuit = steerer.find_feature(
